pyproject.toml
[tool.poetry]
name = "covidfaq"
version = "0.0.0"
description = ""
authors = ["Dialogue <devs@dialogue.co>"]

[tool.poetry.dependencies]
python = "^3.7"
rasa = {extras = ["spacy"], version = "^1.8.2"}
tensorflow-estimator = "2.1.0"
elasticsearch = "^7.6.0"
spacy_langdetect = "^0.1.2"
structlog = "^20.1.0"
beautifulsoup4 = "^4.8.2"
datetime = "^4.3"
fastapi = "^0.52.0"
hypercorn = "^0.9.2"
"dialogue.logging" = "^4.0.1"
"dialogue.tracing" = "^1.0.0"
analytics-python = "^1.2.9"
colorama = "^0.4.3"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
isort = "^4.3.21"
pylama = "^7.7.1"

[[tool.poetry.source]]
name = "dialogue"
url = "https://pypi.fury.io/dialogue"
[build-system]
requires = ["poetry>=1.0.0"]
build-backend = "poetry.masonry.api"

